requirements.txt
fastapi[all]
openai
python-dotenv
BeautifulSoup
requests